+++ actix-rt/CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Unreleased - 2021-xx-xx
 
+* Add `System::run_until_stop` to allow retrieving the exit code on stop. [#411]
+
+[#411]: https://github.com/actix/actix-net/pull/411
 
 ## 2.4.0 - 2021-11-05
 * Add `Arbiter::try_current` for situations where thread may or may not have Arbiter context. [#408]